fetch/metadata/generated/speculation-rules-header.sub.tentative.html

<!DOCTYPE html>
<!--
This test was procedurally generated. Please do not modify it directly.
Sources:
- fetch/metadata/tools/fetch-metadata.conf.yml
- fetch/metadata/tools/templates/speculation-rules-header.sub.html
-->
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/fetch/metadata/resources/helper.sub.js"></script>
<body>
<script>
'use strict';

async function induceRequest(url, test) {
  const win = window.open();
  test.add_cleanup(() => win.close());
  win.location = `/common/blank.html?pipe=header(Speculation-Rules,"${encodeURIComponent(url)}")`;
}

const responseParams = {
  mime: 'application/speculationrules+json',
  body: '{}',
};

promise_test((t) => {
  const key = '{{uuid()}}';

  return induceRequest(
      makeRequestURL(key, ['httpOrigin'], responseParams), t
    )
    .then(() => retrieve(key, {poll: true}))
    .then((headers) => {
        assert_not_own_property(headers, 'sec-fetch-site');
      });
}, 'sec-fetch-site - Not sent to non-trustworthy same-origin destination');

promise_test((t) => {
  const key = '{{uuid()}}';

  return induceRequest(
      makeRequestURL(key, ['httpSameSite'], responseParams), t
    )
    .then(() => retrieve(key, {poll: true}))
    .then((headers) => {
        assert_not_own_property(headers, 'sec-fetch-site');
      });
}, 'sec-fetch-site - Not sent to non-trustworthy same-site destination');

promise_test((t) => {
  const key = '{{uuid()}}';

  return induceRequest(
      makeRequestURL(key, ['httpCrossSite'], responseParams), t
    )
    .then(() => retrieve(key, {poll: true}))
    .then((headers) => {
        assert_not_own_property(headers, 'sec-fetch-site');
      });
}, 'sec-fetch-site - Not sent to non-trustworthy cross-site destination');

promise_test((t) => {
  const key = '{{uuid()}}';

  return induceRequest(
      makeRequestURL(key, ['httpOrigin'], responseParams), t
    )
    .then(() => retrieve(key, {poll: true}))
    .then((headers) => {
        assert_not_own_property(headers, 'sec-fetch-mode');
      });
}, 'sec-fetch-mode - Not sent to non-trustworthy same-origin destination');

promise_test((t) => {
  const key = '{{uuid()}}';

  return induceRequest(
      makeRequestURL(key, ['httpSameSite'], responseParams), t
    )
    .then(() => retrieve(key, {poll: true}))
    .then((headers) => {
        assert_not_own_property(headers, 'sec-fetch-mode');
      });
}, 'sec-fetch-mode - Not sent to non-trustworthy same-site destination');

promise_test((t) => {
  const key = '{{uuid()}}';

  return induceRequest(
      makeRequestURL(key, ['httpCrossSite'], responseParams), t
    )
    .then(() => retrieve(key, {poll: true}))
    .then((headers) => {
        assert_not_own_property(headers, 'sec-fetch-mode');
      });
}, 'sec-fetch-mode - Not sent to non-trustworthy cross-site destination');

promise_test((t) => {
  const key = '{{uuid()}}';

  return induceRequest(
      makeRequestURL(key, ['httpOrigin'], responseParams), t
    )
    .then(() => retrieve(key, {poll: true}))
    .then((headers) => {
        assert_not_own_property(headers, 'sec-fetch-dest');
      });
}, 'sec-fetch-dest - Not sent to non-trustworthy same-origin destination');

promise_test((t) => {
  const key = '{{uuid()}}';

  return induceRequest(
      makeRequestURL(key, ['httpSameSite'], responseParams), t
    )
    .then(() => retrieve(key, {poll: true}))
    .then((headers) => {
        assert_not_own_property(headers, 'sec-fetch-dest');
      });
}, 'sec-fetch-dest - Not sent to non-trustworthy same-site destination');

promise_test((t) => {
  const key = '{{uuid()}}';

  return induceRequest(
      makeRequestURL(key, ['httpCrossSite'], responseParams), t
    )
    .then(() => retrieve(key, {poll: true}))
    .then((headers) => {
        assert_not_own_property(headers, 'sec-fetch-dest');
      });
}, 'sec-fetch-dest - Not sent to non-trustworthy cross-site destination');

promise_test((t) => {
  const key = '{{uuid()}}';

  return induceRequest(
      makeRequestURL(key, ['httpOrigin'], responseParams), t
    )
    .then(() => retrieve(key, {poll: true}))
    .then((headers) => {
        assert_not_own_property(headers, 'sec-fetch-user');
      });
}, 'sec-fetch-user - Not sent to non-trustworthy same-origin destination');

promise_test((t) => {
  const key = '{{uuid()}}';

  return induceRequest(
      makeRequestURL(key, ['httpSameSite'], responseParams), t
    )
    .then(() => retrieve(key, {poll: true}))
    .then((headers) => {
        assert_not_own_property(headers, 'sec-fetch-user');
      });
}, 'sec-fetch-user - Not sent to non-trustworthy same-site destination');

promise_test((t) => {
  const key = '{{uuid()}}';

  return induceRequest(
      makeRequestURL(key, ['httpCrossSite'], responseParams), t
    )
    .then(() => retrieve(key, {poll: true}))
    .then((headers) => {
        assert_not_own_property(headers, 'sec-fetch-user');
      });
}, 'sec-fetch-user - Not sent to non-trustworthy cross-site destination');

promise_test((t) => {
  const key = '{{uuid()}}';

  return induceRequest(
      makeRequestURL(key, ['httpsOrigin', 'httpOrigin'], responseParams), t
    )
    .then(() => retrieve(key, {poll: true}))
    .then((headers) => {
        assert_not_own_property(headers, 'sec-fetch-site');
      });
}, 'sec-fetch-site - HTTPS downgrade (header not sent)');

promise_test((t) => {
  const key = '{{uuid()}}';

  return induceRequest(
      makeRequestURL(key, ['httpOrigin', 'httpsOrigin'], responseParams), t
    )
    .then(() => retrieve(key, {poll: true}))
    .then((headers) => {
        assert_own_property(headers, 'sec-fetch-site');
        assert_array_equals(headers['sec-fetch-site'], ['cross-site']);
      });
}, 'sec-fetch-site - HTTPS upgrade');

promise_test((t) => {
  const key = '{{uuid()}}';

  return induceRequest(
      makeRequestURL(key, ['httpsOrigin', 'httpOrigin', 'httpsOrigin'], responseParams), t
    )
    .then(() => retrieve(key, {poll: true}))
    .then((headers) => {
        assert_own_property(headers, 'sec-fetch-site');
        assert_array_equals(headers['sec-fetch-site'], ['cross-site']);
      });
}, 'sec-fetch-site - HTTPS downgrade-upgrade');
</script>
